refactor(registry): split registry client modules
