+++ crates/aube-registry/src/client.rs
@@ -1,74 +1,18 @@
 use crate::config::{FetchPolicy, NpmConfig};
 use crate::{Error, NetworkMode, Packument};
-use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 use std::sync::Mutex;
 
-/// Disk-cached packument with revalidation metadata.
-#[derive(Debug, Serialize, Deserialize)]
-struct CachedPackument {
-    etag: Option<String>,
-    last_modified: Option<String>,
-    /// Unix epoch seconds when this entry was written
-    fetched_at: u64,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    max_age_secs: Option<u64>,
-    packument: Packument,
-}
+mod body;
+mod cache;
+mod http;
+mod tarball;
 
-/// Disk-cached *full* (non-corgi) packument. Stored as raw JSON so we
-/// preserve fields the resolver doesn't parse (`description`, `repository`,
-/// `license`, `keywords`, `maintainers`, ...), for use by human-facing
-/// commands like `aube view`.
-#[derive(Debug, Serialize, Deserialize)]
-struct CachedFullPackument {
-    etag: Option<String>,
-    last_modified: Option<String>,
-    fetched_at: u64,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    max_age_secs: Option<u64>,
-    packument: serde_json::Value,
-}
-
-#[derive(Debug, Default)]
-pub struct CachedPackumentLookup {
-    pub packument: Option<Packument>,
-    pub stale: bool,
-    cached: Option<CachedPackumentLookupEntry>,
-}
-
-#[derive(Debug)]
-enum CachedPackumentLookupEntry {
-    Abbreviated(CachedPackument),
-    Full(CachedFullPackumentTyped),
-}
-
-#[derive(Debug)]
-struct CachedFullPackumentTyped {
-    etag: Option<String>,
-    last_modified: Option<String>,
-    fetched_at: u64,
-    max_age_secs: Option<u64>,
-    packument: Packument,
-}
-
-fn cached_is_fresh(fetched_at: u64, max_age_secs: Option<u64>) -> bool {
-    let age = now_secs().saturating_sub(fetched_at);
-    let budget = max_age_secs.unwrap_or(PACKUMENT_TTL_SECS);
-    age < budget
-}
-
-/// How long to trust a cached packument before revalidating with the registry.
-/// Trust cached packuments for 30 minutes before revalidating. This keeps
-/// repeated installs in a long-lived dev session from devolving into hundreds
-/// of conditional metadata requests once the cache is just over pnpm's 5-minute
-/// default staleness window.
-const PACKUMENT_TTL_SECS: u64 = 1800;
-
-fn is_retriable_status(status: reqwest::StatusCode) -> bool {
-    status.is_server_error() || status == reqwest::StatusCode::TOO_MANY_REQUESTS
-}
+use body::*;
+pub use cache::CachedPackumentLookup;
+use cache::*;
+use http::*;
 
 /// Accept header for packument requests. `vnd.npm.install-v1+json` is the
 /// abbreviated (corgi) format npmjs emits for installs; the `application/json`
@@ -96,54 +40,6 @@ const PACKUMENT_FULL_ACCEPT: &str = "application/json; q=1.0, */*";
 /// extremely generous upper bound for monorepos with tens of thousands
 /// of locked versions.
 const AUDIT_BODY_CAP: u64 = 256 << 20;
-
-fn now_secs() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0)
-}
-
-/// Pull ETag + Last-Modified off a response as owned strings.
-fn extract_cache_headers(resp: &reqwest::Response) -> (Option<String>, Option<String>) {
-    let headers = resp.headers();
-    let grab = |name: reqwest::header::HeaderName| {
-        headers
-            .get(name)
-            .and_then(|v| v.to_str().ok())
-            .map(|s| s.to_string())
-    };
-    (
-        grab(reqwest::header::ETAG),
-        grab(reqwest::header::LAST_MODIFIED),
-    )
-}
-
-fn parse_cache_control_max_age(resp: &reqwest::Response) -> Option<u64> {
-    let raw = resp
-        .headers()
-        .get(reqwest::header::CACHE_CONTROL)
-        .and_then(|v| v.to_str().ok())?;
-    let mut max_age = None;
-    let mut s_maxage = None;
-    let mut force_revalidate = false;
-    for directive in raw.split(',').map(str::trim) {
-        let directive_lc = directive.to_ascii_lowercase();
-        match directive_lc.as_str() {
-            "no-store" | "no-cache" | "private" => force_revalidate = true,
-            _ => {}
-        }
-        if let Some(val) = directive_lc.strip_prefix("s-maxage=") {
-            s_maxage = val.parse::<u64>().ok();
-        } else if let Some(val) = directive_lc.strip_prefix("max-age=") {
-            max_age = val.parse::<u64>().ok();
-        }
-    }
-    if force_revalidate {
-        return Some(0);
-    }
-    s_maxage.or(max_age)
-}
 
 /// Client for interacting with the npm registry.
 pub struct RegistryClient {
@@ -1723,119 +1619,6 @@ impl RegistryClient {
         Ok(json)
     }
 
-    /// Download a tarball and return the bytes.
-    ///
-    /// Emits a `fetchMinSpeedKiBps` warning when the end-to-end average
-    /// throughput of the body read falls below the configured threshold.
-    /// Average (not instantaneous) speed because the call path is a
-    /// single `resp.bytes().await?` — we keep the eager-read model and
-    /// still give operators a signal for flaky links. `fetchWarnTimeoutMs`
-    /// does *not* fire here: that one is scoped to metadata requests
-    /// per its pnpm documentation, and the tarball-specific analogue
-    /// is the min-speed warning.
-    pub async fn fetch_tarball_bytes(&self, url: &str) -> Result<bytes::Bytes, Error> {
-        // Refuse non-http(s) tarball URLs at the aube boundary so
-        // attacker-controlled `dist.tarball` from a hostile mirror
-        // cannot reach `file:///` (local file disclosure) or the
-        // ssh / git transports inside reqwest. Belt-and-suspenders
-        // against transport-layer regressions.
-        let safe_url = aube_util::url::redact_url(url);
-        let parsed = reqwest::Url::parse(url)
-            .map_err(|e| Error::Io(std::io::Error::other(format!("invalid tarball url: {e}"))))?;
-        match parsed.scheme() {
-            "https" | "http" => {}
-            scheme => {
-                return Err(Error::Io(std::io::Error::other(format!(
-                    "tarball {safe_url}: refusing scheme {scheme:?}",
-                ))));
-            }
-        }
-        if self.network_mode == NetworkMode::Offline {
-            return Err(Error::Offline(format!("tarball {safe_url}")));
-        }
-        // Tarball URLs may point to any registry, try to match auth.
-        // Pass the full tarball URL through so longest-prefix matching
-        // in `registry_config_for` can find path-scoped auth entries
-        // (e.g. `//host/artifactory/npm/`). Tarballs are already gzip
-        // archives, so ask intermediaries not to wrap them in HTTP
-        // content encoding that can fail independently of the payload.
-        // Retries cover transient 5xx / 429 / connection errors; see
-        // [`Self::send_with_retry`].
-        let (bytes, body_elapsed) = self
-            .retry_bytes_body_read(url, self.fetch_policy.tarball_max_bytes, || {
-                self.authed_tarball_get(url, url)
-                    .header(reqwest::header::ACCEPT_ENCODING, "identity")
-            })
-            .await?;
-        warn_slow_tarball(
-            self.fetch_policy.min_speed_kibps,
-            url,
-            bytes.len(),
-            body_elapsed,
-        );
-        Ok(bytes)
-    }
-
-    /// Streaming variant of `fetch_tarball_bytes`. Returns the body
-    /// bytes plus the SHA-512 of the on-the-wire payload, computed
-    /// incrementally during the chunk read loop. Callers that already
-    /// know the lockfile-pinned `integrity` field can compare against
-    /// this digest directly and skip the second hash pass that
-    /// `aube_store::verify_integrity` would otherwise do over the
-    /// owned `Bytes`.
-    ///
-    /// `AUBE_DISABLE_STREAMING_SHA512=1` is the killswitch: callers
-    /// can short-circuit to `fetch_tarball_bytes` and re-hash on the
-    /// import side. The killswitch lives in the caller (so it can pick
-    /// the buffered path without still paying the streaming cost), not
-    /// in this method.
-    pub async fn fetch_tarball_bytes_streaming_sha512(
-        &self,
-        url: &str,
-    ) -> Result<(bytes::Bytes, [u8; 64]), Error> {
-        let _diag = aube_util::diag::Span::new(
-            aube_util::diag::Category::Registry,
-            "tarball_buffered_with_sha512",
-        )
-        .with_meta_fn(|| {
-            format!(
-                r#"{{"url":{}}}"#,
-                aube_util::diag::jstr(&aube_util::url::redact_url(url))
-            )
-        });
-        let safe_url = aube_util::url::redact_url(url);
-        let parsed = reqwest::Url::parse(url)
-            .map_err(|e| Error::Io(std::io::Error::other(format!("invalid tarball url: {e}"))))?;
-        match parsed.scheme() {
-            "https" | "http" => {}
-            scheme => {
-                return Err(Error::Io(std::io::Error::other(format!(
-                    "tarball {safe_url}: refusing scheme {scheme:?}",
-                ))));
-            }
-        }
-        if self.network_mode == NetworkMode::Offline {
-            return Err(Error::Offline(format!("tarball {safe_url}")));
-        }
-        let (bytes, sha512, body_elapsed) = self
-            .retry_bytes_body_read_streaming_sha512(
-                url,
-                self.fetch_policy.tarball_max_bytes,
-                || {
-                    self.authed_tarball_get(url, url)
-                        .header(reqwest::header::ACCEPT_ENCODING, "identity")
-                },
-            )
-            .await?;
-        warn_slow_tarball(
-            self.fetch_policy.min_speed_kibps,
-            url,
-            bytes.len(),
-            body_elapsed,
-        );
-        Ok((bytes, sha512))
-    }
-
     /// Fetch a single VersionMetadata via the per-version registry
     /// endpoint `{registry}/{name}/{version}`. Returns ~1-4 KiB JSON
     /// vs the full packument's 100 KiB-2 MiB. Use when caller knows
@@ -1865,111 +1648,6 @@ impl RegistryClient {
             "version-metadata",
         )?;
         parse_full_response(resp).await
-    }
-
-    /// Start a streaming tarball fetch. Returns the live reqwest
-    /// Response so the caller can pull `chunk()` futures and pipe them
-    /// through gz+tar+CAS without buffering the full body.
-    ///
-    /// Retries the *initial* request on transient failures (5xx, 429,
-    /// connection errors) using `fetch_policy.retries` attempts with
-    /// exponential backoff. Once chunks start flowing the caller owns
-    /// stream-level errors — restarting mid-body would require
-    /// unwinding partial CAS writes — so the caller should fall back
-    /// to `fetch_tarball_bytes_streaming_sha512` (which retries the
-    /// full body cleanly via a buffered fetch) if a mid-stream error
-    /// needs another attempt.
-    pub async fn start_tarball_stream(&self, url: &str) -> Result<reqwest::Response, Error> {
-        let _diag =
-            aube_util::diag::Span::new(aube_util::diag::Category::Registry, "tarball_stream_open")
-                .with_meta_fn(|| {
-                    format!(
-                        r#"{{"url":{}}}"#,
-                        aube_util::diag::jstr(&aube_util::url::redact_url(url))
-                    )
-                });
-        let safe_url = aube_util::url::redact_url(url);
-        let parsed = reqwest::Url::parse(url)
-            .map_err(|e| Error::Io(std::io::Error::other(format!("invalid tarball url: {e}"))))?;
-        match parsed.scheme() {
-            "https" | "http" => {}
-            scheme => {
-                return Err(Error::Io(std::io::Error::other(format!(
-                    "tarball {safe_url}: refusing scheme {scheme:?}",
-                ))));
-            }
-        }
-        if self.network_mode == NetworkMode::Offline {
-            return Err(Error::Offline(format!("tarball {safe_url}")));
-        }
-
-        let label = format!("tarball {safe_url}");
-        let max_attempts = self.fetch_policy.retries.saturating_add(1);
-        let mut timeout_retries: u32 = 0;
-        for attempt in 0..max_attempts {
-            let is_last = attempt + 1 >= max_attempts;
-            let result = self
-                .authed_tarball_get(url, url)
-                .header(reqwest::header::ACCEPT_ENCODING, "identity")
-                .send()
-                .await;
-            match result {
-                Ok(resp) if is_retriable_status(resp.status()) && !is_last => {
-                    let wait = retry_after_from(&resp)
-                        .unwrap_or_else(|| self.fetch_policy.backoff_for_attempt(attempt + 1));
-                    tracing::warn!(
-                        attempt = attempt + 1,
-                        max_attempts,
-                        backoff_ms = wait.as_millis() as u64,
-                        status = resp.status().as_u16(),
-                        label = label.as_str(),
-                        code = aube_codes::warnings::WARN_AUBE_HTTP_RETRY_TRANSIENT,
-                        "retrying HTTP request after transient failure",
-                    );
-                    drop(resp);
-                    tokio::time::sleep(wait).await;
-                }
-                Ok(resp) => {
-                    let resp = resp.error_for_status()?;
-                    check_body_cap(&resp, self.fetch_policy.tarball_max_bytes, "tarball")?;
-                    return Ok(resp);
-                }
-                Err(err) if !is_last => {
-                    if err.is_timeout() {
-                        if timeout_retries >= TIMEOUT_RETRY_CAP {
-                            return Err(Error::Http(err));
-                        }
-                        timeout_retries += 1;
-                    }
-                    let wait = self.fetch_policy.backoff_for_attempt(attempt + 1);
-                    tracing::warn!(
-                        attempt = attempt + 1,
-                        max_attempts,
-                        backoff_ms = wait.as_millis() as u64,
-                        error = %err,
-                        label = label.as_str(),
-                        code = aube_codes::warnings::WARN_AUBE_HTTP_RETRY_TRANSPORT,
-                        "retrying HTTP request after transport error",
-                    );
-                    tokio::time::sleep(wait).await;
-                }
-                Err(err) => return Err(Error::Http(err)),
-            }
-        }
-        // FetchPolicy::retries is `u32`, so `max_attempts =
-        // retries + 1` is always ≥ 1 and the loop runs at least once;
-        // every path inside the loop either returns or continues. An
-        // exit past this point is a structural bug, not a runtime
-        // input the caller can provoke.
-        unreachable!("retry loop exited without returning; max_attempts was {max_attempts}")
-    }
-
-    /// Tarball body cap, surfaced so the streaming caller can enforce
-    /// it as chunks arrive (Content-Length pre-check happens in
-    /// start_tarball_stream but chunked-encoding bodies need ongoing
-    /// total tracking).
-    pub fn tarball_max_bytes(&self) -> u64 {
-        self.fetch_policy.tarball_max_bytes
     }
 
     /// Fetch the *full* (non-corgi) packument as raw JSON, bypassing the
@@ -2136,399 +1814,6 @@ impl Default for RegistryClient {
     }
 }
 
-/// Add inline PEM strings and a PEM-bundle file to a reqwest client
-/// builder as additional trust roots. Shared between the top-level
-/// (unscoped) and per-registry cert paths so both go through the same
-/// parse + warn pipeline.
-fn apply_extra_root_certs(
-    mut builder: reqwest::ClientBuilder,
-    ca: &[String],
-    cafile: Option<&Path>,
-    scope: &str,
-) -> reqwest::ClientBuilder {
-    for pem in ca {
-        match reqwest::Certificate::from_pem(pem.as_bytes()) {
-            Ok(cert) => builder = builder.add_root_certificate(cert),
-            Err(e) => tracing::warn!(
-                code = aube_codes::warnings::WARN_AUBE_INVALID_CA,
-                "ignoring invalid {scope} ca: {e}"
-            ),
-        }
-    }
-    if let Some(cafile) = cafile {
-        match std::fs::read(cafile) {
-            Ok(bytes) => match reqwest::Certificate::from_pem_bundle(&bytes) {
-                Ok(certs) => {
-                    for cert in certs {
-                        builder = builder.add_root_certificate(cert);
-                    }
-                }
-                Err(e) => tracing::warn!(
-                    code = aube_codes::warnings::WARN_AUBE_INVALID_CAFILE,
-                    "ignoring invalid {scope} cafile {}: {e}",
-                    cafile.display()
-                ),
-            },
-            Err(e) => tracing::warn!(
-                code = aube_codes::warnings::WARN_AUBE_UNREADABLE_CAFILE,
-                "ignoring unreadable {scope} cafile {}: {e}",
-                cafile.display()
-            ),
-        }
-    }
-    builder
-}
-
-fn build_http_client(
-    config: &NpmConfig,
-    registry_config: Option<&crate::config::AuthConfig>,
-    fetch_policy: &FetchPolicy,
-) -> reqwest::Client {
-    build_http_client_inner(config, registry_config, fetch_policy, false)
-}
-
-/// HTTP/1.1-only variant for tarball downloads. Tarballs are large
-/// opaque blobs where h2 multiplexing buys nothing: there are no
-/// compressible headers, and a single slow tarball stream causes
-/// head-of-line blocking for every other in-flight tarball on the
-/// same h2 connection. npm's CDN advertises
-/// `SETTINGS_MAX_CONCURRENT_STREAMS` ≈ 100-128, so a 256-permit
-/// tarball semaphore over a single h2 connection queues 128+
-/// requests inside hyper waiting for streams. A diag-trace cold
-/// install observed `tarball_stream_open` mean 565ms (n=1230,
-/// 3242ms on critical path) — that's server-side h2 stream
-/// queueing, not TLS or network.
-///
-/// Switching to h1 lets reqwest's connection pool open as many
-/// parallel TCP connections to `registry.npmjs.org` as we have
-/// in-flight tarball requests (capped by `pool_max_idle_per_host`),
-/// matching what npm/pnpm/yarn already do for the same reason.
-/// Packument requests stay on the h2 client because gzip+brotli
-/// header compression and request multiplexing are real wins for
-/// thousands of small JSON payloads.
-fn build_http_tarball_client(
-    config: &NpmConfig,
-    registry_config: Option<&crate::config::AuthConfig>,
-    fetch_policy: &FetchPolicy,
-) -> reqwest::Client {
-    build_http_client_inner(config, registry_config, fetch_policy, true)
-}
-
-fn build_http_client_inner(
-    config: &NpmConfig,
-    registry_config: Option<&crate::config::AuthConfig>,
-    fetch_policy: &FetchPolicy,
-    for_tarball: bool,
-) -> reqwest::Client {
-    // `maxsockets` (when set) overrides the default pool size. pnpm
-    // documents this as "concurrent connections per origin"; reqwest
-    // doesn't expose a hard cap, but `pool_max_idle_per_host` is the
-    // closest knob and is what downstream users actually care about.
-    let pool_max_idle = config.max_sockets.unwrap_or(64);
-    // CDN edge cache hit rate keys partly off the User-Agent header.
-    // Hardcoded `0.1.0` lands in cold buckets on Cloudflare/Fastly. Use
-    // the real workspace version + an OS/arch tail in the same shape
-    // pnpm and npm send so the registry recognises us.
-    static UA: std::sync::OnceLock<String> = std::sync::OnceLock::new();
-    let user_agent = UA.get_or_init(|| {
-        format!(
-            "aube/{} ({} {})",
-            env!("CARGO_PKG_VERSION"),
-            std::env::consts::OS,
-            std::env::consts::ARCH
-        )
-    });
-    let mut builder = reqwest::Client::builder()
-        .user_agent(user_agent)
-        // Wire-level decompression for packument JSON. Tarball
-        // requests explicitly send `Accept-Encoding: identity`
-        // (tarballs are already gzip on the payload), so this only
-        // affects metadata calls. Popular packuments (`react`,
-        // `webpack`, `next`) drop 3-5x on the wire when gzipped.
-        .gzip(true)
-        .brotli(true)
-        .zstd(true)
-        // `fetchTimeout` — applied to the whole response (headers +
-        // body) via reqwest's single-knob timeout. pnpm / npm expose
-        // this as `fetch-timeout` in `.npmrc`; the default matches
-        // npm's 60s. Without this override reqwest would use its
-        // built-in 30s default, which is tighter than pnpm's.
-        .timeout(std::time::Duration::from_millis(fetch_policy.timeout_ms))
-        // Bigger connection pool so concurrent fetches don't queue on a small set of conns.
-        // HTTP/2 (when negotiated via ALPN, which npm registry supports) multiplexes many
-        // requests over a single connection so this mostly matters for fallback HTTP/1.1.
-        .pool_max_idle_per_host(pool_max_idle)
-        .pool_idle_timeout(std::time::Duration::from_secs(90))
-        .tcp_nodelay(true)
-        // Apply h2 transport tuning only for non-tarball clients.
-        // Tarball client gates below to h1; the h2 setters would be
-        // accepted but never exercised, so we skip them entirely.
-    ;
-    if !for_tarball {
-        builder = builder
-            .http2_keep_alive_interval(std::time::Duration::from_secs(30))
-            .http2_keep_alive_timeout(std::time::Duration::from_secs(20))
-            .http2_keep_alive_while_idle(true)
-            .http2_adaptive_window(true)
-            .http2_initial_stream_window_size(Some(16 * 1024 * 1024))
-            .http2_initial_connection_window_size(Some(16 * 1024 * 1024))
-            .http2_max_frame_size(Some(16 * 1024 * 1024 - 1));
-    } else {
-        builder = builder.http1_only();
-    }
-    builder = builder
-        .tcp_keepalive(std::time::Duration::from_secs(60))
-        // In-process DNS caching via hickory-dns. The system resolver
-        // does not cache and uses a thread pool for `getaddrinfo`,
-        // which serializes the first cold lookup per origin. hickory
-        // resolves async + caches for the process lifetime.
-        .hickory_dns(true)
-        // `strict-ssl=false` disables cert validation entirely. This
-        // is a security hole on purpose: corporate registries should
-        // prefer per-registry `ca` / `cafile` so validation stays on.
-        .danger_accept_invalid_certs(!config.strict_ssl)
-        // rustls already defaults to TLS 1.2+, but pinning the floor
-        // here makes the policy explicit so a future default-loosening
-        // upstream does not silently re-enable TLS 1.1 for aube.
-        .min_tls_version(reqwest::tls::Version::TLS_1_2)
-        // Block https to http downgrades on redirect. reqwest already
-        // strips Authorization on cross-host redirects as of 0.12, so
-        // this policy only adds the scheme guard. A 302 from a good
-        // registry to `http://evil/` would otherwise leak whatever
-        // header survived into cleartext.
-        .redirect(reqwest::redirect::Policy::custom(|attempt| {
-            if attempt.previous().len() >= 10 {
-                return attempt.error("too many redirects");
-            }
-            if let Some(prev) = attempt.previous().last()
-                && prev.scheme() == "https"
-                && attempt.url().scheme() != "https"
-            {
-                return attempt.stop();
-            }
-            attempt.follow()
-        }))
-        // Disable reqwest's built-in `system-proxy` auto-detection
-        // before installing any explicit proxies. Without this, the
-        // builder would silently read `HTTP(S)_PROXY` / `NO_PROXY`
-        // from the environment *on top of* the values we already
-        // pulled into `NpmConfig`, so a `.npmrc` that overrides an
-        // env-var proxy would be ignored for one scheme and honored
-        // for the other, and `noproxy` bypasses would only apply to
-        // the manually-configured proxies. `NpmConfig::load` now
-        // folds the env vars into the config itself, so this crate
-        // is the single source of truth for proxy state.
-        .no_proxy();
-
-    if let Some(ip) = config.local_address {
-        builder = builder.local_address(Some(ip));
-    }
-
-    let no_proxy = config
-        .no_proxy
-        .as_deref()
-        .and_then(reqwest::NoProxy::from_string);
-
-    if let Some(ref url) = config.https_proxy {
-        match reqwest::Proxy::https(url) {
-            Ok(mut p) => {
-                if let Some(ref np) = no_proxy {
-                    p = p.no_proxy(Some(np.clone()));
-                }
-                builder = builder.proxy(p);
-            }
-            Err(e) => tracing::warn!(
-                code = aube_codes::warnings::WARN_AUBE_INVALID_HTTPS_PROXY,
-                "ignoring https-proxy {url:?}: {e}"
-            ),
-        }
-    }
-    if let Some(ref url) = config.http_proxy {
-        match reqwest::Proxy::http(url) {
-            Ok(mut p) => {
-                if let Some(ref np) = no_proxy {
-                    p = p.no_proxy(Some(np.clone()));
-                }
-                builder = builder.proxy(p);
-            }
-            Err(e) => tracing::warn!(
-                code = aube_codes::warnings::WARN_AUBE_INVALID_HTTP_PROXY,
-                "ignoring http-proxy {url:?}: {e}"
-            ),
-        }
-    }
-
-    // Top-level `cafile` / `ca` (unscoped npmrc keys) apply to every
-    // client built from this config, matching npm/pnpm semantics.
-    builder = apply_extra_root_certs(builder, &config.ca, config.cafile.as_deref(), "top-level");
-
-    if let Some(registry_config) = registry_config {
-        builder = apply_extra_root_certs(
-            builder,
-            &registry_config.tls.ca,
-            registry_config.tls.cafile.as_deref(),
-            "per-registry",
-        );
-        if let (Some(cert), Some(key)) = (&registry_config.tls.cert, &registry_config.tls.key) {
-            let mut pem = Vec::with_capacity(cert.len() + key.len() + 1);
-            pem.extend_from_slice(cert.as_bytes());
-            if !cert.ends_with('\n') {
-                pem.push(b'\n');
-            }
-            pem.extend_from_slice(key.as_bytes());
-            match reqwest::Identity::from_pem(&pem) {
-                Ok(identity) => builder = builder.identity(identity),
-                Err(e) => tracing::warn!(
-                    code = aube_codes::warnings::WARN_AUBE_INVALID_CLIENT_CERT,
-                    "ignoring invalid per-registry client cert/key: {e}"
-                ),
-            }
-        }
-    }
-
-    builder.build().expect("failed to build HTTP client")
-}
-
-/// BATS-fixture escape hatch: ask the registry for the unabbreviated
-/// packument instead of the corgi (`application/vnd.npm.install-v1+json`)
-/// shape. Our Verdaccio-backed fixture strips `bundledDependencies`
-/// when it projects stored packuments to corgi, so the
-/// `test/bundled_dependencies.bats` suite sets this to exercise the
-/// resolver's bundled-skip path end-to-end. Production registries
-/// include `bundleDependencies` in corgi per the npm spec, so the
-/// default path stays cheap.
-///
-/// The name is deliberately `AUBE_INTERNAL_*` so nothing outside the
-/// test harness grows a habit of relying on it, and we require the
-/// exact literal `"1"` (not just any non-empty value) so an inherited
-/// or accidentally-set empty value won't silently balloon registry
-/// traffic on end-user machines.
-fn force_full_packument() -> bool {
-    std::env::var("AUBE_INTERNAL_FORCE_FULL_PACKUMENT").as_deref() == Ok("1")
-}
-
-/// Refuse a response whose declared `Content-Length` exceeds `cap`
-/// before reading the body. A hostile registry (or MITM on a
-/// compromised mirror) could otherwise stream gigabytes into the
-/// resolver and OOM the install. Servers that omit `Content-Length`
-/// (chunked transfer) still reach `bytes()` below, where the read is
-/// bounded by the caller's operational timeout. The full-streaming
-/// cap-while-reading variant is left for a follow-up, since it needs
-/// a `futures` / `tokio-stream` dep that the crate does not yet pull
-/// in.
-///
-/// A `cap` of `0` disables the check entirely — an escape hatch for
-/// users who need to pull packuments that exceed the default (e.g.
-/// packages with very long release histories) and accept the DoS
-/// exposure on the trusted-registry side.
-/// Stream-and-count read of a response body that enforces `cap` even
-/// when the server omits `Content-Length` (chunked transfer encoding).
-/// `check_body_cap` only inspects the precheck header; this function
-/// is the runtime gate that closes the chunked-bypass primitive.
-async fn read_body_capped(
-    mut resp: reqwest::Response,
-    cap: u64,
-    label: &str,
-) -> Result<bytes::Bytes, Error> {
-    if cap == 0 {
-        return Ok(resp.bytes().await?);
-    }
-    const STREAM_INITIAL: usize = 64 * 1024;
-    let initial = resp
-        .content_length()
-        .map(|len| len.min(cap) as usize)
-        .unwrap_or(STREAM_INITIAL);
-    let mut buf = bytes::BytesMut::with_capacity(initial);
-    while let Some(chunk) = resp.chunk().await? {
-        if (buf.len() as u64).saturating_add(chunk.len() as u64) > cap {
-            return Err(Error::Io(std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                format!("{label}: response body exceeds cap {cap}"),
-            )));
-        }
-        buf.extend_from_slice(&chunk);
-    }
-    Ok(buf.freeze())
-}
-
-/// `read_body_capped` plus a streaming SHA-512 of every byte the
-/// registry delivered. Used by the tarball fetch path to skip the
-/// post-buffer integrity hash (~7 ms / 5 MB tarball, ~50-120 ms /
-/// 1000-pkg cold install). `Accept-Encoding: identity` is set on
-/// the tarball request so reqwest cannot transparently decompress
-/// underneath us, which means the streamed digest covers the same
-/// bytes `aube_store::verify_integrity` checks against the lockfile
-/// `integrity` field. Non-tarball callers go through the buffered
-/// `read_body_capped` and skip the per-chunk hash work.
-async fn read_body_capped_streaming_sha512(
-    mut resp: reqwest::Response,
-    cap: u64,
-    label: &str,
-) -> Result<(bytes::Bytes, [u8; 64]), Error> {
-    use sha2::Digest;
-    const STREAM_INITIAL: usize = 64 * 1024;
-    // Pre-size from Content-Length when present, capped at `cap`
-    // when set, falling back to a 64 KiB scratch when neither is
-    // available so chunked-encoding bodies don't pay BytesMut's
-    // doubling-grow tax all the way up to `cap`.
-    let initial = match (resp.content_length(), cap) {
-        (Some(len), 0) => len as usize,
-        (Some(len), cap) => len.min(cap) as usize,
-        (None, _) => STREAM_INITIAL,
-    };
-    let mut buf = bytes::BytesMut::with_capacity(initial);
-    let mut hasher = sha2::Sha512::new();
-    while let Some(chunk) = resp.chunk().await? {
-        if cap > 0 && (buf.len() as u64).saturating_add(chunk.len() as u64) > cap {
-            return Err(Error::Io(std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                format!("{label}: response body exceeds cap {cap}"),
-            )));
-        }
-        hasher.update(&chunk);
-        buf.extend_from_slice(&chunk);
-    }
-    let mut digest = [0u8; 64];
-    digest.copy_from_slice(&hasher.finalize()[..]);
-    Ok((buf.freeze(), digest))
-}
-
-fn check_body_cap(resp: &reqwest::Response, cap: u64, label: &str) -> Result<(), Error> {
-    if cap == 0 {
-        return Ok(());
-    }
-    if let Some(len) = resp.content_length()
-        && len > cap
-    {
-        return Err(Error::Io(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            format!("{label}: response Content-Length {len} exceeds cap {cap}"),
-        )));
-    }
-    Ok(())
-}
-
-fn packument_cache_path(cache_dir: &Path, name: &str, registry_url: &str) -> Option<PathBuf> {
-    // `name` is derived from registry responses and user-written
-    // manifests. `replace('/', "__")` alone would let `../../evil`
-    // escape the cache directory and turn a first resolve into an
-    // arbitrary-file-write primitive. Delegate to the store's
-    // shared validator so the grammar never drifts across crates.
-    let safe_name = aube_store::validate_and_encode_name(name)?;
-    // Partition by registry origin: a packument fetched against
-    // registry A must never be returned to a request that resolves
-    // to registry B (CVE-2018-7167 class). Hash the URL so port,
-    // trailing-slash, and scheme variants share the same bucket only
-    // when literally identical bytes were configured.
-    let origin = registry_origin_segment(registry_url);
-    Some(cache_dir.join(origin).join(format!("{safe_name}.json")))
-}
-
-fn registry_origin_segment(registry_url: &str) -> String {
-    let digest = blake3::hash(registry_url.as_bytes()).to_hex();
-    format!("origin-{}", &digest.as_str()[..16])
-}
-
 /// URL-encode a package name for the `/-/package/<name>/...` path.
 /// Only `/` needs encoding — scoped packages have exactly one, between
 /// `@scope` and `pkg`. npm expects `@scope%2Fpkg` for scoped names on
@@ -2603,188 +1888,6 @@ where
     );
     result
 }
-
-fn read_cached_packument(path: &Path) -> Option<CachedPackument> {
-    // sonic-rs is faster than serde_json on packument-shape JSON and,
-    // unlike simd-json, takes an immutable `&[u8]` so the file content
-    // doesn't need to be kept mutable for the parse to be zero-copy.
-    let content = std::fs::read(path).ok()?;
-    sonic_rs::from_slice(&content).ok()
-}
-
-fn write_cached_packument(path: &Path, cached: &CachedPackument) -> std::io::Result<()> {
-    // sonic-rs serializer for symmetry with the read path; output
-    // format doesn't need to match anything external (cache file we
-    // own), so we trade serde_json's stable formatting for a small
-    // throughput win on the cold-install metadata phase.
-    let json = sonic_rs::to_vec(cached).map_err(std::io::Error::other)?;
-    aube_util::fs_atomic::atomic_write(path, &json)
-}
-
-fn packument_full_cache_path(cache_dir: &Path, name: &str, registry_url: &str) -> Option<PathBuf> {
-    let safe_name = aube_store::validate_and_encode_name(name)?;
-    let origin = registry_origin_segment(registry_url);
-    Some(cache_dir.join(origin).join(format!("{safe_name}.json")))
-}
-
-fn read_cached_full_packument(path: &Path) -> Option<CachedFullPackument> {
-    let content = std::fs::read(path).ok()?;
-    sonic_rs::from_slice(&content).ok()
-}
-
-/// Typed fast-path read used by `fetch_packument_with_time_cached`
-/// in the warm-cache branch. Reads the file once and uses `sonic-rs`
-/// to deserialize the cached wrapper directly into a tiny typed struct
-/// holding `fetched_at` plus a fully-typed [`Packument`].
-///
-/// Returns a missing lookup on file/parse errors, and a stale lookup
-/// when revalidation is needed, so callers can decide whether a primer
-/// fallback is safe without reading the cache a second time.
-fn read_cached_full_packument_typed_lookup(
-    path: &Path,
-    force_cache: bool,
-) -> CachedPackumentLookup {
-    #[derive(Deserialize)]
-    struct Typed {
-        etag: Option<String>,
-        last_modified: Option<String>,
-        fetched_at: u64,
-        #[serde(default)]
-        max_age_secs: Option<u64>,
-        packument: Packument,
-    }
-
-    let Ok(content) = std::fs::read(path) else {
-        return CachedPackumentLookup::default();
-    };
-    let Ok(typed) = sonic_rs::from_slice::<Typed>(&content) else {
-        return CachedPackumentLookup::default();
-    };
-    let typed = CachedFullPackumentTyped {
-        etag: typed.etag,
-        last_modified: typed.last_modified,
-        fetched_at: typed.fetched_at,
-        max_age_secs: typed.max_age_secs,
-        packument: typed.packument,
-    };
-    if !force_cache && !cached_is_fresh(typed.fetched_at, typed.max_age_secs) {
-        return CachedPackumentLookup {
-            packument: None,
-            stale: true,
-            cached: Some(CachedPackumentLookupEntry::Full(typed)),
-        };
-    }
-    CachedPackumentLookup {
-        packument: Some(typed.packument),
-        stale: false,
-        cached: None,
-    }
-}
-
-fn read_cached_full_packument_typed(path: &Path, force_cache: bool) -> Option<Packument> {
-    read_cached_full_packument_typed_lookup(path, force_cache).packument
-}
-
-fn write_cached_full_packument(
-    path: &Path,
-    etag: Option<&str>,
-    last_modified: Option<&str>,
-    fetched_at: u64,
-    max_age_secs: Option<u64>,
-    packument: &serde_json::Value,
-) -> std::io::Result<()> {
-    // Serialize through a borrow struct so popular packuments don't pay
-    // a multi-MB `serde_json::Value::clone` per write. The owned
-    // `CachedFullPackument` is still used by the read path; the writer
-    // just doesn't need ownership.
-    #[derive(Serialize)]
-    struct CachedFullPackumentRef<'a> {
-        etag: Option<&'a str>,
-        last_modified: Option<&'a str>,
-        fetched_at: u64,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        max_age_secs: Option<u64>,
-        packument: &'a serde_json::Value,
-    }
-    let json = sonic_rs::to_vec(&CachedFullPackumentRef {
-        etag,
-        last_modified,
-        fetched_at,
-        max_age_secs,
-        packument,
-    })
-    .map_err(std::io::Error::other)?;
-    aube_util::fs_atomic::atomic_write(path, &json)
-}
-
-/// Emit a `fetchMinSpeedKiBps` warning if the tarball downloaded slower
-/// than the configured threshold. `threshold_kibps == 0` disables the
-/// warning (pnpm convention). Transfers that completed in one second
-/// or less are skipped: for small/fast responses the TCP/TLS handshake
-/// and TTFB dominate the "average" throughput, producing spurious
-/// warnings that don't reflect network health. This matches pnpm's
-/// `elapsedSec > 1` gate in its tarball fetcher.
-fn warn_slow_tarball(threshold_kibps: u64, url: &str, len: usize, elapsed: std::time::Duration) {
-    if threshold_kibps == 0 {
-        return;
-    }
-    if len == 0 || elapsed <= std::time::Duration::from_secs(1) {
-        return;
-    }
-    let elapsed_ms = elapsed.as_millis() as u64;
-    // speed (KiB/s) = bytes / 1024 / seconds = bytes * 1000 / elapsed_ms / 1024
-    let kibps = ((len as u64).saturating_mul(1000)) / elapsed_ms / 1024;
-    if kibps < threshold_kibps {
-        let safe_url = aube_util::url::redact_url(url);
-        tracing::warn!(
-            kibps,
-            threshold_kibps,
-            bytes = len,
-            elapsed_ms,
-            url = %safe_url,
-            code = aube_codes::warnings::WARN_AUBE_SLOW_TARBALL,
-            "slow tarball download fell below fetchMinSpeedKiBps",
-        );
-    }
-}
-
-/// Parse the `Retry-After` response header as a number of seconds.
-/// Per RFC 7231, this header can also be an HTTP-date, but the `Date`
-/// format is rare in practice for npm-style registries and `chrono`
-/// isn't a dep — callers fall back to the computed exponential
-/// backoff if the header is missing, unparseable, or in date form.
-/// `RETRY_AFTER_CAP_SECS` clamps the parsed value so a hostile
-/// registry can't park an install for hours or years by returning
-/// `Retry-After: 999999999`.
-fn retry_after_from(resp: &reqwest::Response) -> Option<std::time::Duration> {
-    let raw = resp
-        .headers()
-        .get(reqwest::header::RETRY_AFTER)?
-        .to_str()
-        .ok()?;
-    let secs: u64 = raw.trim().parse().ok()?;
-    Some(std::time::Duration::from_secs(
-        secs.min(RETRY_AFTER_CAP_SECS),
-    ))
-}
-
-/// Upper bound on `Retry-After` we are willing to honour. 60 seconds
-/// is well above any real npm-style rate-limit cooldown and keeps the
-/// total retry budget bounded even when a server hands us a bogus
-/// value.
-const RETRY_AFTER_CAP_SECS: u64 = 60;
-
-/// Maximum number of timeout-shaped retries before we surface the
-/// error to the caller, regardless of `fetchRetries`. A timeout has
-/// already cost us `fetchTimeout` of wall-clock; retrying many more
-/// times compounds the user-visible hang without much chance of
-/// recovery on the same upstream. One retry is enough to absorb a
-/// fluke; beyond that, fail fast and let the caller decide.
-///
-/// Counted separately from the global retry counter inside the retry
-/// loop so a non-timeout failure (e.g. a 503 on the first attempt)
-/// never consumes the timeout budget.
-const TIMEOUT_RETRY_CAP: u32 = 1;
 
 #[cfg(test)]
 mod seed_tests {
